### 2018-02-08 #576 @arfon
@serine - it looks like there is an issue with your bibtex file. I think the issue is that [this URL](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe/blob/master/paper.bib#L447) needs quoting.

Once you've fixed this you can as @whedon to compile your paper again by typing `@whedon generate pdf`
