### 2018-02-08 #576 @serine
Right, thanks guys, I wasn't sure how to include URL citation in `.bib` let work on this today.

Also what are you building with ? `pandocs`? If I can build with `pandoc` it'll be alright?
