### 2018-02-09 3d92c39
fixed error in .bib file, now compiles with pandoc --bibliography=paper.bib flag
