+++ paper.bib
@@ -444,11 +444,13 @@
     month = September,
     year = 2015,
     publisher = GitHub,
-    url = http://degust.erc.monash.edu
+    url = {http://degust.erc.monash.edu}
 }
 @misc{picard-tools2018,
+    author = "Broadinstitute",
+    title = "Picard-Tools",
     publisher = broadinstitute,
-    url = http://broadinstitute.github.io/picard
+    url = {http://broadinstitute.github.io/picard}
 }
 @ARTICLE{Robinson2011-du,
   title     = "Integrative genomics viewer",
