### 2018-02-09 #576 @arfon
@serine - thanks for debugging this. The PDF is looking better now.
