### 2018-02-12 27a6eab
added mk_igv_links.py script as general utility script
