+++ scripts/mk_igv_links.py
@@ -1,0 +1,167 @@
+#!/usr/bin/env python3
+# -*- coding: iso-8859-15 -*-
+
+import sys
+import os
+import argparse
+
+def str2bool(v):
+    if v.lower() in ('yes', 'true', 't', 'y', '1'):
+        return True
+    elif v.lower() in ('no', 'false', 'f', 'n', '0'):
+        return False
+    else:
+        raise argparse.ArgumentTypeError('Boolean value expected.')
+
+parser = argparse.ArgumentParser(usage='%(prog)s --dir <path/to/your_data>',
+                                 description="scripts for making URL links that you can use to view your data in IGV. Several file types are supported",
+                                 add_help=True
+                                 )
+parser.add_argument('--dir',
+                    required=True,
+                    help="specify directory with your data files to be made into IGV links. This directory should be located in the hostable place i.e you should be able to grab individual file on url"
+                    )
+parser.add_argument('--coord',
+                    help="this will be your default coordinates when IGV loads for the first time when IGV loads. Pass coordinates as a string as follows chr12:24565477-24624103"
+                    )
+parser.add_argument('--host_name',
+                    default='file:///',
+                    help='provide your hosting server name [file:///]'
+                    )
+parser.add_argument('--genome_id',
+                    help="specify your species genome id. Either IGV inbuilt onces or your own, custome made. Note if custome made genome id is used, you'll need to cache .genome file first, see --add_genome_url options for more."
+                    ),
+parser.add_argument('--add_genome_url',
+                    help="specify url to your .genome file, can do file:///"
+                    ),
+parser.add_argument('--igv_meta',
+                    type=str2bool,
+                    nargs='?',
+                    default=True,
+                    help="Add meta information, describes in brief what IGV is and a coupel of links to IGV resources"
+                    )
+parser.add_argument('--sep',
+                    default='_',
+                    help = 'This is main separator within the file'
+                    )
+
+args = parser.parse_args()
+dir = args.dir
+coord = args.coord
+host_name = args.host_name
+genome_id = args.genome_id
+genome_url = args.add_genome_url
+igv_meta = args.igv_meta
+sep = args.sep
+
+if not dir.endswith("/"):
+    dir = dir+'/'
+
+allowed_files = ['bam', 'gtf', 'gff3', 'vcf', 'fa', 'fasta', 'wig', 'bw', 'bedGraph']
+#NOTE this function going to work on our servers, by probably wont elsewhere
+# due to www in the directory path
+def get_url(host_name, var):
+    home_var = os.getenv('HOME')
+    full_path = os.path.abspath(var)
+    my_url = host_name + full_path[full_path.index(home_var):]
+    #my_url = my_url.replace(':', '%3A').replace("/", "%2F")
+    return my_url
+
+#NOTE this is some what a hack to make it work with custome made .genome files
+# basically I'm assuming that custome made .genome file is located in some directory
+# e.g igv_ref/check.genome and I'll pass it in like this --genome igv_refs/check.genome
+# if I would have passed simply --genome check.genome then host_name dose not get prefixed
+# and when you load your data igv assumes that there is a local copy of check.genome
+# whereas with host_name prefix igv will actually download that genome.
+#TODO whole custome .genome files are a little confusing. I think I got main idea.
+# basically if genome=http://host_name/blah.genome, which is a keyword inside url encoding
+# then igv will attemp to download this resource into its local directory, on linux ~/igv/genomes
+# if genome=blah.genome then igv will search ~/igv/genomes for that file, parse and load it genome
[…]
