### 2018-02-12 bcf9507
updated docs to inlcude sample report
