### 2018-02-12 #576 @serine
@pjotrp FYI, I just tweaked the docs a little, essentially added this [link](http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/) with a bit more explanation.

`RNAsik` will return `multiqc_report.html`. The other two links were done in addition to the pipeline run.

- To make [Degust](http://degust.erc.monash.edu) session you'll need upload counts file.

- IGV links were done with `mk_igv_link.py` now located in [`scripts/`](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe/tree/master/scripts) directory

We are working on automatically pushing counts to Degust, but that would require experimental break down i.e grouping samples into conditions, will be done through `samplesSheet.txt` later on.
Can't really do automatic `igv_links.html` because data needs to be in hostable place, i.e you gotta have nginx or apache running on your server, hence running `mk_igv_link.py` manually
