### 2018-02-14 #583 @whedon
[REVIEW]: RNAsik: A Pipeline for complete and reproducible RNA-seq analysis that runs anywhere with speed and ease
**Submitting author:** @serine (<a href="http://orcid.org/0000-0002-9195-2626">Kirill Tsyganov</a>)
**Repository:** <a href="https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe" target ="_blank">https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe</a>
**Version:** 1.4.9
