### 2018-02-15 18b60c3
typo in the docs installation, fixed #15
