### 2018-02-16 MonashBioinformaticsPlatform/RNAsik-pipe#16 @andrewyatz
Cannot install bds for RNAsik-pipe
Hi. Just got the Ansible code running. Everything was fine until I hit the following:

```
TASK [bds : build BigDataScript 0.99999g with java_local] ************************************************************************************************************************
fatal: [localhost]: FAILED! => {"changed": true, "cmd": "cd /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/; ./scripts/install.sh", "delta": "0:00:25.324549", "end": "2018-02-16 14:25:09.317847", "msg": "non-zero return code", "rc": 127, "start": "2018-02-16 14:24:43.993298", "stderr": "+ BDS_HOME=/Users/ayates/.bds\n++ dirname ./scripts/install.sh\n+ echo 'Changing dir' ./scripts\n++ dirname ./scripts/install.sh\n+ cd ./scripts\n+ cd ..\n+ mkdir -p /Users/ayates/.bds\n+ mkdir bin\n+ echo Building JAR file\n+ ant\n+ echo\n+ echo Building GO program\n+ cd go/bds/\n++ pwd\n+ export GOPATH=/Users/ayates/bioansible/software/source/BigDataScript-0.99999g/go/bds\n+ GOPATH=/Users/ayates/bioansible/software/source/BigDataScript-0.99999g/go/bds\n+ go clean\n./scripts/install.sh: line 27: go: command not found", "stderr_lines": ["+ BDS_HOME=/Users/ayates/.bds", "++ dirname ./scripts/install.sh", "+ echo 'Changing dir' ./scripts", "++ dirname ./scripts/install.sh", "+ cd ./scripts", "+ cd ..", "+ mkdir -p /Users/ayates/.bds", "+ mkdir bin", "+ echo Building JAR file", "+ ant", "+ echo", "+ echo Building GO program", "+ cd go/bds/", "++ pwd", "+ export GOPATH=/Users/ayates/bioansible/software/source/BigDataScript-0.99999g/go/bds", "+ GOPATH=/Users/ayates/bioansible/software/source/BigDataScript-0.99999g/go/bds", "+ go clean", "./scripts/install.sh: line 27: go: command not found"], "stdout": "Changing dir ./scripts\nBuilding JAR file\nBuildfile: /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/build.xml\n\nclean:\n   [delete] Deleting directory /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin\n    [mkdir] Created dir: /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin\n\ncompile:\n    [javac] Compiling 403 source files to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin\n    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:9: warning: FtpURLConnection is internal proprietary API and may be removed in a future release\n    [javac] import sun.net.www.protocol.ftp.FtpURLConnection;\n    [javac]                                ^\n    [javac] warning: Supported source version 'RELEASE_6' from annotation processor 'org.antlr.v4.runtime.misc.NullUsageProcessor' less than -source '1.8'\n    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:9: warning: FtpURLConnection is internal proprietary API and may be removed in a future release\n    [javac] import sun.net.www.protocol.ftp.FtpURLConnection;\n    [javac]                                ^\n    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:9: warning: FtpURLConnection is internal proprietary API and may be removed in a future release\n    [javac] import sun.net.www.protocol.ftp.FtpURLConnection;\n    [javac]                                ^\n    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:27: warning: FtpURLConnection is internal proprietary API and may be removed in a future release\n    [javac] \t\t((FtpURLConnection) connection).close();\n    [javac] \t\t  ^\n    [javac] 5 warnings\n\ncreate_run_jar:\n     [copy] Copying 1 file to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin/org/bds\n     [copy] Copying 1 file to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin/org/bds\n     [copy] Copying 1 file to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin/org/bds\n      [jar] Building jar: /Users/ayates/.bds/bds.jar\n\nBUILD SUCCESSFUL\nTotal time: 24 seconds\n\nBuilding GO program", "stdout_lines": ["Changing dir ./scripts", "Building JAR file", "Buildfile: /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/build.xml", "", "clean:", "   [delete] Deleting directory /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin", "    [mkdir] Created dir: /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin", "", "compile:", "    [javac] Compiling 403 source files to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin", "    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:9: warning: FtpURLConnection is internal proprietary API and may be removed in a future release", "    [javac] import sun.net.www.protocol.ftp.FtpURLConnection;", "    [javac]                                ^", "    [javac] warning: Supported source version 'RELEASE_6' from annotation processor 'org.antlr.v4.runtime.misc.NullUsageProcessor' less than -source '1.8'", "    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:9: warning: FtpURLConnection is internal proprietary API and may be removed in a future release", "    [javac] import sun.net.www.protocol.ftp.FtpURLConnection;", "    [javac]                                ^", "    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:9: warning: FtpURLConnection is internal proprietary API and may be removed in a future release", "    [javac] import sun.net.www.protocol.ftp.FtpURLConnection;", "    [javac]                                ^", "    [javac] /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/src/org/bds/data/DataFtp.java:27: warning: FtpURLConnection is internal proprietary API and may be removed in a future release", "    [javac] \t\t((FtpURLConnection) connection).close();", "    [javac] \t\t  ^", "    [javac] 5 warnings", "", "create_run_jar:", "     [copy] Copying 1 file to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin/org/bds", "     [copy] Copying 1 file to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin/org/bds", "     [copy] Copying 1 file to /Users/ayates/bioansible/software/source/BigDataScript-0.99999g/bin/org/bds", "      [jar] Building jar: /Users/ayates/.bds/bds.jar", "", "BUILD SUCCESSFUL", "Total time: 24 seconds", "", "Building GO program"]}
```

Do you know what's gone wrong here?
