### 2018-02-18 MonashBioinformaticsPlatform/RNAsik-pipe#16 @serine
@andrewyatz sorry about late ish replay. So building BigDataScript (BDS) requires three dependencies

- java1.8
- go (go lang)
- ant

From the error it appears that `golang` isn't installed. can you do this? This is from office [github wiki](https://github.com/golang/go/wiki/Ubuntu)

Try:
```
sudo apt-get install golang-go
```
If that's too old for you, try:
```
sudo add-apt-repository ppa:gophers/archive
sudo apt-get update
sudo apt-get install golang-1.9-go
```
This is for debian based distros of course

Thanks for trying
