### 2018-02-18 MonashBioinformaticsPlatform/RNAsik-pipe#16 @serine
Also @andrewyatz once everything has been installed, can you double check that this `bds.config` file has `644` permission on it. I think there is a minor glitch in `bio-ansible` that I need to fix.

That file (`bds.config`) should be located about there  `...apps/BigDataScript-version/bds.config` this will depend on your installation path. Default installation path is `~/bioansible/software/apps/BigDataScript...`

To change permissions use `sudo chmod 644 path/to/your/bds.config`

p.s `bio-ansible` is a separate project that I'm involved in, with multiple developers. I'm just using `bio-ansible` to install `RNAsik` with all its dependencies. All those errors are not directly related to `RNAsik`. Feel free to suggest an alternative build/installation way. I've been considering these two:

- `conda` which isn't just python specific, it is actually quite generic. The problem with `conda` package, we need an existing `BigDataScript` and `GoLang` conda packages, which we currently don't have.

- `docker container` very common approach is to build and provide docker image. We have that in beta, but I'm not a big fan of docker mainly because it isn't as easy to run as people think. You can't run docker images on most HPC clusters due to docker's security model.
