### 2018-02-18 MonashBioinformaticsPlatform/RNAsik-pipe#16 @serine
Just one more comment to clarify `bio-ansible` and `RNAsik` installation, specifically this error with `BDS` installation.

`bio-ansible` would install `golang`, `ant` and all other required dependencies if you run with `all.yml` instead of `bio.yml`, however `all.yml` requires `sudo` privileged and will `apt-get` install stuff for you. I would only recommend using `all.yml` if you are using ephemeral instance, that you gonna scrape after you've tested `RNAsik`.

If you have spare minute have a glance at [bio-ansible docs](https://github.com/MonashBioinformaticsPlatform/bio-ansible)

cheers
