### 2018-02-19 MonashBioinformaticsPlatform/RNAsik-pipe#16 @andrewyatz
Thank you for the details. I'd suggest you should change your docs to say that java1.8, go and ant are requirements. I couldn't find anything in https://monashbioinformaticsplatform.github.io/RNAsik-pipe/docs/ mentioning this requirement. In addition you should add your preferred installation operation system to your docs. Both of these would have been of great help before
