### 2018-02-19 MonashBioinformaticsPlatform/RNAsik-pipe#16 @andrewyatz
Another query:

```
TASK [bds : moving executable bds 0.99999g to apps directory] ******************
fatal: [localhost]: FAILED! => {"changed": true, "cmd": "mv \"/Users/ayates/.bds\" \"/Users/ayates/bioansible/software/apps/BigDataScript-0.99999g\"", "delta": "0:00:00.023850", "end": "2018-02-19 13:03:55.847116", "msg": "non-zero return code", "rc": 1, "start": "2018-02-19 13:03:55.823266", "stderr": "mv: rename /Users/ayates/.bds to /Users/ayates/bioansible/software/apps/BigDataScript-0.99999g: No such file or directory", "stderr_lines": ["mv: rename /Users/ayates/.bds to /Users/ayates/bioansible/software/apps/BigDataScript-0.99999g: No such file or directory"], "stdout": "", "stdout_lines": []}
```

I can confirm I have a `~/bioansible/software` directory but no `apps` directory underneath this.
