### 2018-02-20 d3da821
improved installation docs attemp to fix #16
