+++ docs/docs.md
@@ -86,7 +86,38 @@ As mentioned previously in [about section](index.md#about) very first step in [R
 As one can imagine every one of those tools has several number of options and by running [RNAsik-pipeline](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe) you get predefined - subjective run. Obviously it all comes from years of experience and continues development and improvement. Use can always pass his/her own options through `-extraOptions` flag for more fine turning. 
 Alternatively as, hinted above, user can leverage of [RNAsik](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe) to run everything separately with fine control over the individual run. [RNAsik](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe) produces [.html report](https://en.wikipedia.org/wiki/HTML) with all commands options specified.
 
-## Prerequisites
+## Installation
+
+### Quick start
+
+```
+sudo apt-get install unzip make gcc git python-virtualenv
+sudo apt-get install zlib1g-dev libbz2-dev liblzma-dev libncurses5-dev
+sudo apt-get install openjdk-8-jdk ant golang-go
+sudo apt-get install git htop tmux vim
+
+virtualenv ~/ansible_env
+source ~/ansible_env/bin/activate
+
+pip install --upgrade pip
+pip install ansible
+
+git clone https://github.com/MonashBioinformaticsPlatform/bio-ansible
+cd bio-ansible/
+ansible-playbook -i hosts bio.yml --tags dirs,bds,rnasik,star,bwa,hisat2,subread,samtools,htslib,bedtools,picard,qualimap,fastqc,multiqc
+
+export PATH=$HOME/bioansible/software/apps/BigDataScript-0.99999g:$PATH
+export PATH=$HOME/bioansible/software/apps/RNAsik-pipe-1.4.9/bin:$PATH
+
+RNAsik
+```
+
+### Tools prerequisites
+
+I tried to account for every sub-dependency in [bio-ansible](https://github.com/MonashBioinformaticsPlatform/bio-ansible), definitely checked against vanilla ubuntu 16.04 linux distro, but other systems/linux distros might have slightly deviation from this. If you run into trouble please double check dependencies for the tool that is failing.
+There is quite a spectrum of languages there in the pipeline, C/C++, java and python so far. One can image the difficulty to accommodate every distro and/or system.
+
+p.s doing my best here !
 
 - [BigDataScript](http://pcingola.github.io/BigDataScript/download.html)
 - [STAR aligner](https://github.com/alexdobin/STAR/releases)
@@ -98,20 +129,66 @@ Alternatively as, hinted above, user can leverage of [RNAsik](https://github.com
 - [MultiQC](http://multiqc.info/) 
 - [FastQC](http://www.bioinformatics.babraham.ac.uk/projects/fastqc/)
 
-## Installation
+### System prerequisites
 
-### Preferred method
+In order to install system dependencies you'll need admin privilege i.e `sudo`
+
+general, these are you "stock" utils, that most running system will/should have
+
+- `unzip`
+- `make`
+- `gcc`
+- `git`
+- `python-virtualenv`
+
+```
+sudo apt-get install unzip make gcc git python-virtualenv
+```
+
+samtools, htslib and bwa deps, these are some what specific libraries
+
+- `zlib1g-dev` 
+- `libbz2-dev` 
+- `liblzma-dev`
+- `libncurses5-dev`
+
+```
+sudo apt-get install zlib1g-dev libbz2-dev liblzma-dev libncurses5-dev
+```
+
+Java and BigDataScript, these again rather generic packages, except golang. 
[…]
