### 2018-02-20 f8651cf
spell check and more comments on installation
