+++ docs/docs.md
@@ -114,10 +114,8 @@ RNAsik
 
 ### Tools prerequisites
 
-I tried to account for every sub-dependency in [bio-ansible](https://github.com/MonashBioinformaticsPlatform/bio-ansible), definitely checked against vanilla ubuntu 16.04 linux distro, but other systems/linux distros might have slightly deviation from this. If you run into trouble please double check dependencies for the tool that is failing.
-There is quite a spectrum of languages there in the pipeline, C/C++, java and python so far. One can image the difficulty to accommodate every distro and/or system.
-
-p.s doing my best here !
+I tried to account for every sub-dependency in [bio-ansible](https://github.com/MonashBioinformaticsPlatform/bio-ansible), definitely checked against vanilla ubuntu 16.04 linux distro, however other systems/linux distros might have slight deviation from this. If you run into trouble please double check dependencies for the tool that is failing.
+There is quite a spectrum of languages there in the pipeline, C/C++, java and python so far. One can image the difficulty to accommodate every distro and/or system. I'm doing my best !
 
 - [BigDataScript](http://pcingola.github.io/BigDataScript/download.html)
 - [STAR aligner](https://github.com/alexdobin/STAR/releases)
@@ -133,7 +131,7 @@ p.s doing my best here !
 
 In order to install system dependencies you'll need admin privilege i.e `sudo`
 
-general, these are you "stock" utils, that most running system will/should have
+**General**, these are you "stock" utils, that most running system will/should have
 
 - `unzip`
 - `make`
@@ -145,7 +143,7 @@ general, these are you "stock" utils, that most running system will/should have
 sudo apt-get install unzip make gcc git python-virtualenv
 ```
 
-samtools, htslib and bwa deps, these are some what specific libraries
+**samtools, htslib and bwa deps**, these are some what specific libraries
 
 - `zlib1g-dev` 
 - `libbz2-dev` 
@@ -156,8 +154,8 @@ samtools, htslib and bwa deps, these are some what specific libraries
 sudo apt-get install zlib1g-dev libbz2-dev liblzma-dev libncurses5-dev
 ```
 
-Java and BigDataScript, these again rather generic packages, except golang. 
-Note that golang is pretty easy to install, comes as a pre-compiiled binary [here](https://golang.org/dl/) if you need to
+**Java and BigDataScript**, these again rather generic packages, except golang. 
+Note that golang is pretty easy to install, comes as a pre-compiled binary [here](https://golang.org/dl/) if you don't want to get it through system package mamanger
 
 - `openjdk-8-jdk` 
 - `ant`
@@ -167,8 +165,8 @@ Note that golang is pretty easy to install, comes as a pre-compiiled binary [her
 sudo apt-get install openjdk-8-jdk ant golang-go
 ```
 
-extras, these are optional dependencies, but `tmux` especially recommended as pipeline run could take some time to complete
-_provided you are going to run on remote machine (server), which is also recommended_
+**Extras**, these are optional dependencies, but `tmux` especially recommended as pipeline run could take some time to complete
+_provided you are doing this on remote machine (server), which is also recommended_
 
 - `htop`
 - `tmux`
@@ -201,8 +199,12 @@ path/to/RNAsik-pipe/bin/RNAsik
 
 ### Make RNAsik analysis ready
 
[…]
