### 2018-02-21 83f439d
more spell check and more comments on RNAsik installation
