### 2018-02-21 MonashBioinformaticsPlatform/RNAsik-pipe#16 @serine
@andrewyatz I worked a bit more on the docs, particular installation part. Also made a little youtube video of following the docs installation live :)

Can you please have another go at it?

Just to reflect on you previous comment about golang and and being undocumented. Those are BigDataScript dependencies and so I assumed that those could be found on BigDataScript install page. It is a bit tricky to try to cover every tools dependency. I've included most, common/crucial dependencies into `RNAsik` install docs.

Also it appears that you are trying to run/install this on your Mac? As I've mentioned in the docs you'll need fair bit system resources. For the test data set 16 Gb of RAM and at least 4 cpus.

Cheers
