### 2018-02-21 #583 @serine
@pjotrp and @andrewyatz as I've mention in my [previous comment](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe/issues/16#issuecomment-367302865) `RNAsik` depends on a lot of different tools and I some what assumed that dependencies of every tool don't need to be listed under `RNAsik`. In the case of BigDataScripit it depends on java, golang and ant, given that I said BigDataScript is required I assumed it dependencies.

I since updated the docs to reflect all major/crucial dependencies in the docs, that of course requires a few system installs. I've tested the docs out and it worked for me. If you follow [quick start](https://monashbioinformaticsplatform.github.io/RNAsik-pipe/docs/#installation) `RNAsik` should install. As I also mentioned in the docs now, I've only really tested this on ubuntu 16.04 (also works on 14.04). I don't have a mac so can't really test on that, but you wouldn't want to run it on mac anyway given the resources, unless the data is small of course e.g bacteria.

so like I said docs have been updated with new install instruction.

Let me know how it goes. also if you gonna meet some issues on mac do let me know so that I can accommodate mac users too.

thanks
