### 2018-02-21 #583 @serine
Just to follow up on previous comment above. `virtualenv` isn't strictly necessary if you already have `ansible` installed, you can skip `virtualenv` step. Every time you opening a new shell you'll need to `export` both bds and `RNAsik` again. You can put those two export lines into your `~/.bashrc` if you like. I personally don't like to pollute my `~/.bashrc` with random exports.

A general comment on `bio-ansible`, which I think I already made somewhere, you can change `bio.yml` to `all.ym` and add `--ask-become-pass` which will prompt you to your `sudo` password. This will do bunch of `apt-get` installs (again, ubuntu specific install) following by tools installation. The problem with doing `all.yml` this will install potentially many more `apt-get` packages than you really need for `RNAsik`
