### 2018-02-21 #583 @pjotrp
Complicated tools may have complicated setups ;). One solution would be to provide a Docker container (though I am no fan particularly).  @andrewyatz are up up for trying? Possibly a VM would be the easiest.
