### 2018-02-22 #583 @pjotrp
You may want to look into GNU Guix some day. I don't have any problems with the dependency graph.
