+++ scripts/mk_igv_links.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: iso-8859-15 -*-
 
+from __future__ import print_function
 import sys
 import os
 import argparse
+
 
 def str2bool(v):
     if v.lower() in ('yes', 'true', 't', 'y', '1'):
@@ -13,37 +15,54 @@ def str2bool(v):
     else:
         raise argparse.ArgumentTypeError('Boolean value expected.')
 
-parser = argparse.ArgumentParser(usage='%(prog)s --dir <path/to/your_data>',
-                                 description="scripts for making URL links that you can use to view your data in IGV. Several file types are supported",
-                                 add_help=True
-                                 )
-parser.add_argument('--dir',
-                    required=True,
-                    help="specify directory with your data files to be made into IGV links. This directory should be located in the hostable place i.e you should be able to grab individual file on url"
-                    )
-parser.add_argument('--coord',
-                    help="this will be your default coordinates when IGV loads for the first time when IGV loads. Pass coordinates as a string as follows chr12:24565477-24624103"
-                    )
+
+parser = argparse.ArgumentParser(
+    usage='%(prog)s --dir <path/to/your_data>',
+    description="scripts for making URL links that you can use to view your "
+                "data in IGV. Several file types are supported",
+    add_help=True
+)
+parser.add_argument(
+    '--dir',
+    required=True,
+    help="specify directory with your data files to be made into IGV links. "
+         "This directory should be located in the hostable place i.e you "
+         "should be able to grab individual file on url"
+)
+parser.add_argument(
+    '--coord',
+    help="this will be your default coordinates when IGV loads "
+         "for the first time when IGV loads. Pass coordinates "
+         "as a string as follows chr12:24565477-24624103"
+)
 parser.add_argument('--host_name',
                     default='file:///',
                     help='provide your hosting server name [file:///]'
                     )
-parser.add_argument('--genome_id',
-                    help="specify your species genome id. Either IGV inbuilt onces or your own, custome made. Note if custome made genome id is used, you'll need to cache .genome file first, see --add_genome_url options for more."
-                    ),
-parser.add_argument('--add_genome_url',
-                    help="specify url to your .genome file, can do file:///"
-                    ),
-parser.add_argument('--igv_meta',
-                    type=str2bool,
-                    nargs='?',
-                    default=True,
-                    help="Add meta information, describes in brief what IGV is and a coupel of links to IGV resources"
-                    )
-parser.add_argument('--sep',
-                    default='_',
-                    help = 'This is main separator within the file'
-                    )
+parser.add_argument(
+    '--genome_id',
+    help="specify your species genome id. Either IGV inbuilt "
+         "onces or your own, custome made. Note if custome made genome id is "
+         "used, you'll need to cache .genome file first, see --add_genome_url "
+         "options for more."
+),
+parser.add_argument(
+    '--add_genome_url',
+    help="specify url to your .genome file, can do file:///"
+),
+parser.add_argument(
+    '--igv_meta',
+    type=str2bool,
+    nargs='?',
+    default=True,
+    help="Add meta information, describes in brief what IGV is "
+         "and a coupel of links to IGV resources"
+)
+parser.add_argument(
+    '--sep',
+    default='_',
+    help='This is main separator within the file'
+)
 
 args = parser.parse_args()
 dir = args.dir
[…]

[…]
