### 2018-03-09 a347ee3
Make Python2/3 compatible (mostly with futurize). @serine Reviewed, pushed back two minor syntaxtic things, other all good! Thanks heaps
