+++ scripts/get_geneids.py
@@ -1,57 +1,67 @@
 #!/usr/bin/env python
 # -*- coding: iso-8859-15 -*-
 
+from __future__ import print_function
 import argparse
 import sys
 import os
 import re
 
-parser = argparse.ArgumentParser(usage='%(prog)s --file_type [gff|gtf|saf] --in_file <path/to/your/db_file>',
-                                 description="This script creates three column file Gene ID, Gene Name and Gene Biotype",
-                                 add_help=True
-                                 )
-parser.add_argument('--in_file',
-                    required = True,
-                    help="path to your annotation file"
-                    )
-parser.add_argument('--file_type',
-                    required = True,
-                    help="speficy file type [gff|gtf|saf]"
-                    )
-parser.add_argument('--feature_type',
-                    default="gene",
-                    help="specify feature type, this is for GFF file only, feature type comes from third column in GFF file and is case sensitive, default [gene]"
-                    )
+parser = argparse.ArgumentParser(
+    usage='%(prog)s --file_type [gff|gtf|saf] --in_file <path/to/your/db_file>',
+    description="This script creates three column file: "
+                "Gene ID, Gene Name and Gene Biotype",
+    add_help=True
+)
+parser.add_argument(
+    '--in_file',
+    required=True,
+    help="path to your annotation file"
+)
+parser.add_argument(
+    '--file_type',
+    required=True,
+    help="specify file type [gff|gtf|saf]"
+)
+
+parser.add_argument(
+    '--feature_type',
+    default="gene",
+    help="specify feature type, this is for GFF file only, feature type comes from third column in "
+         "GFF file and is case sensitive, default [gene]"
+)
+
 
 def get_gtf(handler):
-
     genes_attr = {}
     typesRegex = "\s.([A-z0-9 _. \- /\(\)]+)"
-    
-    for key, value in biotypes.items():
+
+    for key, value in list(biotypes.items()):
         tweak = value + typesRegex
-        biotypes[key]=tweak
-        
+        biotypes[key] = tweak
+
     for i in handler:
         if not i.startswith("#"):
             line = i.split('\t')
             ninthField = line[8]
 
-            chk_gene_id = re.search('gene_id\s.([A-z0-9_ \- \(\)]+)', ninthField)
-            chk_gene_name = re.search('gene_name\s.([A-z0-9_ \. \: \- \(\)]+)', ninthField)
-    
+            chk_gene_id = re.search('gene_id\s.([A-z0-9_ \- \(\)]+)',
+                                    ninthField)
+            chk_gene_name = re.search('gene_name\s.([A-z0-9_ \. \: \- \(\)]+)',
+                                      ninthField)
+
             gene_name = 'NA'
             gene_biotype = 'NA'
             chrom = line[0]
-    
+
             if chk_gene_name:
                 gene_name = chk_gene_name.group(1)
-    
-            for value in biotypes.values():
+
+            for value in list(biotypes.values()):
                 checkBiotype = re.search(value, ninthField)
                 if checkBiotype:
                     gene_biotype = checkBiotype.group(1)
-    
+
             if chrom not in genes_attr:
                 genes_attr[chrom] = {}
 
@@ -60,29 +70,30 @@ def get_gtf(handler):
 
                 if gene_id not in genes_attr[chrom]:
                     genes_attr[chrom][gene_id] = {}
-            #TODO I feel that this is error prone..
-            # I'mm imaginig a situations where previous gene_name had proper value
-            # but in the next iteration (next line) gene_name tag doesn't exists
-            # and therefor gets NA value and thats now overides proper values already
-            # stored in the dictionary
-            # in theory every line in gtf files is self contained and this shouldn't happened
[…]
