### 2018-03-11 3cda6c2
fixed bug in python script and added new arg option, making it more flexible
