### 2018-03-16 MonashBioinformaticsPlatform/RNAsik-pipe#16 @andrewyatz
Have to reopen this I'm sorry. So I've just tried to run this on vagrant/virtualbox using Ubuntu 16.04 as suggested. Firstly I think your docs should be updated to first issue a `sudo apt-get update` ensuring apt is ready to install. Secondly you should add `--yes` to your install commands as it will make it a lot easier to copy and paste the commands to run them. That's the good news.

Bad news. I got to running bioansible as prescribed on the ubuntu VM and get the following output

```
$ ansible-playbook -i hosts bio.yml --tags dirs,bds,rnasik,star,bwa,hisat2,subread,samtools,htslib,bedtools,picard,qualimap,fastqc,multiqc
[DEPRECATION WARNING]: The use of 'include' for tasks has been deprecated. Use 'import_tasks' for static inclusions or 'include_tasks' for dynamic inclusions. This feature will be removed in a future release. Deprecation warnings can be
 disabled by setting deprecation_warnings=False in ansible.cfg.
[DEPRECATION WARNING]: include is kept for backwards compatibility but usage is discouraged. The module documentation details page may explain more about this rationale.. This feature will be removed in a future release. Deprecation
warnings can be disabled by setting deprecation_warnings=False in ansible.cfg.
[DEPRECATION WARNING]: Specifying include variables at the top-level of the task is deprecated. Please see: http://docs.ansible.com/ansible/playbooks_roles.html#task-include-files-and-encouraging-reuse   for currently supported syntax
regarding included files and variables. This feature will be removed in version 2.7. Deprecation warnings can be disabled by setting deprecation_warnings=False in ansible.cfg.

PLAY [Installing bioinformatics tools] ******************************************************************************************************************************************************************************************************

TASK [Gathering Facts] **********************************************************************************************************************************************************************************************************************
ok: [localhost]

TASK [make_dirs : Create required directories] **********************************************************************************************************************************************************************************************
failed: [localhost] (item=/home/vagrant/bioansible/software/source) => {"changed": false, "gid": 1000, "group": "vagrant", "item": "/home/vagrant/bioansible/software/source", "mode": "0775", "msg": "chown failed: failed to look up user ubuntu", "owner": "vagrant", "path": "/home/vagrant/bioansible", "size": 4096, "state": "directory", "uid": 1000}
failed: [localhost] (item=/home/vagrant/bioansible/software/modules) => {"changed": false, "gid": 1000, "group": "vagrant", "item": "/home/vagrant/bioansible/software/modules", "mode": "0775", "msg": "chown failed: failed to look up user ubuntu", "owner": "vagrant", "path": "/home/vagrant/bioansible/software", "size": 4096, "state": "directory", "uid": 1000}
failed: [localhost] (item=/home/vagrant/bioansible/software/modules/core) => {"changed": false, "gid": 1000, "group": "vagrant", "item": "/home/vagrant/bioansible/software/modules/core", "mode": "0775", "msg": "chown failed: failed to look up user ubuntu", "owner": "vagrant", "path": "/home/vagrant/bioansible/software/modules", "size": 4096, "state": "directory", "uid": 1000}
failed: [localhost] (item=/home/vagrant/bioansible/software/modules/bio) => {"changed": false, "gid": 1000, "group": "vagrant", "item": "/home/vagrant/bioansible/software/modules/bio", "mode": "0775", "msg": "chown failed: failed to look up user ubuntu", "owner": "vagrant", "path": "/home/vagrant/bioansible/software/modules/bio", "size": 4096, "state": "directory", "uid": 1000}
failed: [localhost] (item=/home/vagrant/bioansible/software/modules/extra) => {"changed": false, "gid": 1000, "group": "vagrant", "item": "/home/vagrant/bioansible/software/modules/extra", "mode": "0775", "msg": "chown failed: failed to look up user ubuntu", "owner": "vagrant", "path": "/home/vagrant/bioansible/software/modules/extra", "size": 4096, "state": "directory", "uid": 1000}
failed: [localhost] (item=/home/vagrant/bioansible/software/apps) => {"changed": false, "gid": 1000, "group": "vagrant", "item": "/home/vagrant/bioansible/software/apps", "mode": "0775", "msg": "chown failed: failed to look up user ubuntu", "owner": "vagrant", "path": "/home/vagrant/bioansible/software/apps", "size": 4096, "state": "directory", "uid": 1000}
	to retry, use: --limit @/home/vagrant/bio-ansible/bio.retry

PLAY RECAP **********************************************************************************************************************************************************************************************************************************
localhost                  : ok=1    changed=0    unreachable=0    failed=1
```
Honestly I don't know what is wrong here. There must be an issue in how these directories are being created but I don't know what that is. Please can you take a look.
