### 2018-03-16 #583 @andrewyatz
Okay I've resurrected my work and switched to an Ubuntu box. Still no joy. In fact the same error I see on my mac. This is in Mon
