### 2018-03-19 b2d287d
updated python regex
