### 2018-04-13 712eec2
updated changelog and roadmap with 1.5.0 release
