### 2018-04-17 MonashBioinformaticsPlatform/RNAsik-pipe#22 @serine
@andrewyatz,  looks like your username on that VM is `vagrant`. could you set that in `group_vars/all` under `main_guy: vagrant`? This should solve your problem.

This is a bit annoying it should auto detect localhost username, but for some reason I'm struggling to do that through ansible, hopefully will be able to fix in the future
