### 2018-04-17 #583 @serine
@andrewyatz sorry about that, I've gone quite, my fault. I just responded to your [other issue](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe/issues/22) could you have one more try. Thanks
