### 2018-04-19 684ba84
Updated docs, added conda installation. Moved bio-ansible installation into "Advanced" tab
