+++ docs/docs.md
@@ -5,6 +5,11 @@
 
 ## Quick start
 
+### Install
+
+conda install -c serine rnasik 
+conda install -c bioconda qualimap 
+
 ### Align raw reads
 
 ```BASH
@@ -26,8 +31,7 @@ RNAsik -fqDir /path/to/raw-data/directory \
        -align star \
        -refFiles /path/to/refDir \
        -counts \
-       -metrics \
-       -threads 10
+       -all
 ```
 
 ## Data set for testing
@@ -67,9 +71,8 @@ RNAsik -align star \
        -fqDir http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData.tar \
        -counts \
        -gtfFile ftp://ftp.ensembl.org/pub/release-91/gtf/saccharomyces_cerevisiae/Saccharomyces_cerevisiae.R64-1-1.91.gtf.gz \
-       -metrics \
-       -paired \
-       -threads 10
+       -all \
+       -paired
 ```
 
 ## Introduction
@@ -89,148 +92,33 @@ Alternatively as, hinted above, user can leverage of [RNAsik](https://github.com
 
 ## Installation
 
-[watch it on youtube](https://www.youtube.com/watch?v=oHRa9iniy3o)
+### Using conda
 
-### Quick start
-
-```
-sudo apt-get install unzip make gcc git python-virtualenv
-sudo apt-get install zlib1g-dev libbz2-dev liblzma-dev libncurses5-dev
-sudo apt-get install openjdk-8-jdk ant golang-go
-sudo apt-get install git htop tmux vim
-
-virtualenv ~/ansible_env
-source ~/ansible_env/bin/activate
-
-pip install --upgrade pip
-pip install ansible
-
-git clone https://github.com/MonashBioinformaticsPlatform/bio-ansible
-cd bio-ansible/
-ansible-playbook -i hosts bio.yml --tags dirs,bds,rnasik,star,bwa,hisat2,subread,samtools,htslib,bedtools,picard,qualimap,fastqc,multiqc
-
-export PATH=$HOME/bioansible/software/apps/BigDataScript-0.99999g:$PATH
-export PATH=$HOME/bioansible/software/apps/RNAsik-pipe-1.4.9/bin:$PATH
-
-RNAsik
-```
-
-### Tools prerequisites
-
-I tried to account for every sub-dependency in [bio-ansible](https://github.com/MonashBioinformaticsPlatform/bio-ansible), definitely checked against vanilla ubuntu 16.04 linux distro, however other systems/linux distros might have slight deviation from this. If you run into trouble please double check dependencies for the tool that is failing.
-There is quite a spectrum of languages there in the pipeline, C/C++, java and python so far. One can image the difficulty to accommodate every distro and/or system. I'm doing my best !
-
-- [BigDataScript](http://pcingola.github.io/BigDataScript/download.html)
-- [STAR aligner](https://github.com/alexdobin/STAR/releases)
-- [subread](http://subread.sourceforge.net/)
-- [samtools](http://www.htslib.org/download/)
-- [bedtools2](http://bedtools.readthedocs.io/en/latest/index.html)
-- [Picard tools](http://broadinstitute.github.io/picard/)
-- [QualiMap](http://qualimap.bioinfo.cipf.es/)
-- [MultiQC](http://multiqc.info/) 
-- [FastQC](http://www.bioinformatics.babraham.ac.uk/projects/fastqc/)
-
-### System prerequisites
-
-In order to install system dependencies you'll need admin privilege i.e `sudo`
-
-**General**, these are you "stock" utils, that most running system will/should have
-
[…]
