+++ mkdocs.yml
@@ -10,6 +10,7 @@ pages:
   - Help: help.md
   - Contributing: contrib.md
   - Roadmap: roadmap.md
+  - Advance: install.md
 
 extra_css:
   - css/style.css
