### 2018-04-19 f9e69d2
removed duplicated lines from install.md
