### 2018-04-19 #583 @serine
@andrewyatz It does appear that bio-ansible way to install `RNAsik` is a bit too much. Taking advice from @pjotrp and others around me I've created conda package, for linux only at this stage, which should simplify install great deal.

Could you have a look at [updated docs](https://monashbioinformaticsplatform.github.io/RNAsik-pipe/docs/#installation)

Basically you'll need to install miniconda and then RNAsik from my "channel" (conda terminology here)

```
bash Miniconda3-latest-Linux-x86_64.sh
conda install -c serine rnasik
conda install -c bioconda qualimap
```

`QualiMap` is separate install, because it itself has fair few dependencies and it was tricky to include into `RNAsik` package

On my laptop it took about 35 minutes to install, this is subject to reasonably good internet connection as installation creates fair bit of traffic, downloading all of the different dependencies.

Hopefully this will simplify things for you and others

Cheers
