### 2018-04-20 4f66b94
updated docs, notes on adding conda channels
