### 2018-04-20 #583 @serine
@andrewyatz sorry, I think you also need to added a couple of channel

```
conda config --add channels defaults
conda config --add channels conda-forge
conda config --add channels bioconda
```

before installing from `conda`

also updated docs.

p.s default location of the config is `~/.condarc`

mine looks like this

```
channels:
  - bioconda
  - conda-forge
  - defaults
```
