### 2018-04-23 37fda4b
tweaked get_genesids regex to be more sensitive
