### 2018-05-08 75e4129
tweaked get_geneidx.py script
