### 2018-05-08 010ae85
Added another regex into global multiqc_conf for handling hisat2 output log files
