+++ configs/multiqc_conf.yml
@@ -2,3 +2,4 @@
 
 extra_fn_clean_exts:
     - _sorted
+    - _aligned
