### 2018-05-09 2cabd08
Added description of python scripts used in the pipeline and how to use them manually
