+++ docs/docs.md
@@ -273,6 +273,7 @@ Pipelines runs `featureCounts` three times forcing reads to forward strand only,
 
 <table>
 <tr><th>Options</th><th> Usage </th></tr>
+<tr><td class="left_col">-refFiles</td> <td> path to refFiles/ directory previously generated by RNAsik run</td></tr>
 <tr><td class="left_col">-mdups</td> <td> flag to mark duplicates</td></tr>
 <tr><td class="left_col">-trim</td> <td> perform FASTQ adapter and quality trimming</td></tr>
 <tr><td class="left_col">-cov</td> <td> get coverage plots, bigWig </td></tr>
@@ -300,3 +301,130 @@ Pipelines runs `featureCounts` three times forcing reads to forward strand only,
   href="https://twitter.com/intent/tweet?text=Hey%20I%27m%20using%20this%20fully%20sick%20RNAseq%20pipeline%20It%27s%20sik%20easy%20http%3A%2F%2Fgithub%2Ecom%2Fmonashbioinformaticsplatform%2FRNAsik%2Dpipe%20by%20%40kizza%5Fa%20from%20%40MonashBioinfo" data-size="large">
 Share</a>
 </p>
+
+## Python scripts
+
+There a few python scripts that are used in the pipeline to do various things, mostly to do with read counts post processing.
+Hopefully script name are self explanatory or have a good hint to what they do. All script have help menu, simply run the scripts
+with `-h|--help` to get more information. All script also output to stdout (print to the screen) in order to save the output you'll
+need to redirect to a file with `>` symbol.
+All script set to be executable so you should be able just run them. If for some reason you cannot run them just use `python` prefix
+in front of the script name. Here are some examples on how to run python scripts
+
+```BASH
+./get_geneids.py
+```
+
+OR
+
+```BASH
+/full_path/to/get_geneids.py
+```
+
+OR
+
+```BASH
+python get_geneids.py
+```
+
+- `get_geneids.py`
+
+Converts GTF|GFF files into four columns, tab delimited file `Gene.Id\tChrom\tGene.Name\tBiotype`
+Note that `Gene.Name` and `Biotype` are subject to availability, i.e if that information isn't present
+in your annotation file then script (obviously) can't parse it out.
+We need this information to augment our counts file. This isn't essential for DGE but makes
+our exploratory analysis easier.
+
+- `mk_cnts_file.py`
+
+Once we have our counts files from `featureCounts` and we've created `geneIds.txt` file with previous command
+We can add that extra meta information to our counts file.
+Note that this script filter on biotype, default is [protein_coding]. Also note that if biotype information is
+absent then your `-withNames-proteinCoding.txt` file will be empty since no such biotype was found.
+You can use `--biotype all` to get unfiltered - all genes. This is usually located in `-withNames.txt` file
+
+- `strand_guessing.py`
+
+This script takes three different counts files i.e Forward, Reverse and Non stranded counts and attempts to
+guess which strand the data is. We can get this information from the sequencing facility and most Illumina
+library preparation kits are reverse stranded these days, but sometimes things happened and this approach
+gives us correct answers from the data itself.
+The output of this script is two values, comma separated. e.g `ReverseStandedCounts,0`
+The first value is obvious, tells you the type of strand it guessed. The second value is what's called an
+exit code, zero == all good. It means the script is pretty confident that it guessed it right. The other
+possible value could be 1, 1 == not good. It means the script couldn't guess what the strand was and simply
+defaulted to `NonStrandedCounts`. You should never see exit code 1 with anything else by non stranded counts
+i.e `NonStrandedCounts,1`. If you do, please report this as a bug.
+
+- `mk_igv_links.py`
+
+This script isn't used in the pipeline actually, but it comes rather handy if you want to visualise your
[…]
