### 2018-05-09 b36d2ad
Found unwanted behaviour (bug?) in get_geneids.py, fixed?
