### 2018-05-10 b3f91ca
Refactored strand_guessing.py a little to make it more pythonic Added another field to return statement, now I'm getting three pieces of informaiton, StrandType,StrandValue,ExitCode. Strand value is used to guess strand. It is useful to have a look at that value.
Propagated those changes in RNAsik code base

Also updated docs to note new output from strand_guessing.py
