### 2018-05-10 a9b24d4
Added release note for upcoming 1.5.1
