### 2018-05-11 958fdf8
In relation to 1aac1a3 commit, forgot to apply the same fix to mk_cnts_file.py script. Also needed to sort list of string by length and reverse it, such that logest string - sample name comes first.
