### 2018-05-14 09d7955
added notes, how to build a conda package
