### 2018-05-14 29af7d3
added install-with-conda and anaconda version badge
