+++ README.md
@@ -1,5 +1,8 @@
 # RNAsik-pipe
 
+![install-with-conda](https://anaconda.org/serine/rnasik/badges/installer/conda.svg)
+![anaconda-version](https://anaconda.org/serine/rnasik/badges/version.svg)
+
 #### This is fully sick way to get your differentially expressed genes !
 
 > `RNAsik-pipe` implemented in [BigDataScrip](http://pcingola.github.io/BigDataScript/) (BDS) language.
