### 2018-05-14 MonashBioinformaticsPlatform/RNAsik-pipe#22 @serine
@andrewyatz going to close this issue for now, feel free to reopen it again if you need to. As I've mentioned on the other channel, you can now simply install with `conda install` [more info here](https://monashbioinformaticsplatform.github.io/RNAsik-pipe/docs/#installation)
