### 2018-05-14 4b2b50e
Added links to actual install packge and files listing; anaconda badge
