+++ README.md
@@ -1,7 +1,7 @@
 # RNAsik-pipe
 
-![install-with-conda](https://anaconda.org/serine/rnasik/badges/installer/conda.svg)
-![anaconda-version](https://anaconda.org/serine/rnasik/badges/version.svg)
+[![install-with-conda](https://anaconda.org/serine/rnasik/badges/installer/conda.svg)](https://anaconda.org/serine/rnasik)
+[![anaconda-version](https://anaconda.org/serine/rnasik/badges/version.svg)](https://anaconda.org/serine/rnasik/files)
 
 #### This is fully sick way to get your differentially expressed genes !
 
