### 2018-05-15 e90a73c
Updated multiqc_conf.yml file, fine turning columns order and display
