### 2018-05-24 2893037
Made a couple of improvements to python script, now get_gff properly searches for biotype and script is tolerant to empty or less 9 fields lines
