### 2018-05-29 2b215b5
Updated docs, added travis CI logo, but really I just want to see if travis CI works
