### 2018-05-29 fc507a4
updated travis script
