+++ .travis.yml
@@ -5,12 +5,10 @@ sudo: false
 language: bds
 
 before_install:
-  - if [ ${TRAVIS_PYTHON_VERSION:0:1} == "2" ]
-    then
+  - if [ ${TRAVIS_PYTHON_VERSION:0:1} == "2" ]; then
       wget http://repo.continuum.io/miniconda/Miniconda-2.0.0-Linux-x86_64.sh -O miniconda.sh
     else
-      wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
-    fi
+      wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh; fi
   - bash miniconda.sh -b
   - export PATH=/home/travis/anaconda/bin:$PATH
 
