+++ .travis.yml
@@ -5,15 +5,18 @@ sudo: false
 language: bds
 
 before_install:
+
   - if [[ ${TRAVIS_PYTHON_VERSION:0:1} == "2" ]]; then
-      wget http://repo.continuum.io/miniconda/Miniconda-2.0.0-Linux-x86_64.sh -O miniconda.sh
+      wget http://repo.continuum.io/miniconda/Miniconda-2.0.0-Linux-x86_64.sh -O miniconda.sh;
     else
       wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh; fi
-  - bash miniconda.sh -b
-  - export PATH=/home/travis/anaconda/bin:$PATH
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - export PATH=$HOME/miniconda/bin:$PATH
 
 install:
-  - conda install -c serine bigdatascript
+
+  - conda install --yes --channel serine bigdatascript
 
 script:
-  - bds -test test/sikTest.bds
+
+  - bds -test test/sikTests.bds
