### 2018-05-29 7bd6a76
travis4, added simple test for RNAsik install
