### 2018-05-29 5137e74
travis5, adding conda channels
