+++ .travis.yml
@@ -15,11 +15,16 @@ before_install:
 
 install:
 
+  - conda config --add channels defaults
+  - conda config --add channels conda-forge
+  - conda config --add channels bioconda
+
   # RNAsik should pull bds with it
   #- conda install --yes --channel serine bigdatascript
   - conda install --yes --channel serine rnasik
 
+
 script:
 
   - bds -test test/sikTests.bds
-  - RNAsik
+  - RNAsik --help
