### 2018-05-30 0518d4c
updated docs, added into about testing and travis ci. Also updated roadmap
