+++ docs/contrib.md
@@ -9,17 +9,42 @@ and of course with the source itself.
 ## Documentations
 
 I'm using [mkdocs](https://github.com/mkdocs/mkdocs) to generate this site, which has been very easy to use.
-All documentations are written in plain markdown and located in main repo [`docs/` directory](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe/tree/master/docs). You can simply fork `RNAsik` repository, do appropriate changes to the docs and send me a pull request (PR). Any small changes are super welcomed, even one letter spell correction (there'll be more than one), but all changes need to come through PR, which will not only acknowledge you as contributor, but also enable me to review changes quickly and incorporate them in (pull them in) easily.
+All documentations are written in plain markdown and located in main repo [`docs/` directory](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe/tree/master/docs). You can simply fork `RNAsik` repository, do changes to the docs and send me a pull request (PR). Any changes are super welcomed, even one letter spell correction (there'll be more than one), but all changes need to come through PR, which will not only acknowledge you as contributor, but also enable me to review changes quickly and incorporate them in (pull them in) easily.
 
 Quick notes on [mkdocs](https://github.com/mkdocs/mkdocs), it is pretty easy to install with `pip` in `virtualenv` if you prefer (you should).
+
+- to install mkdocs (don't have to use `virtualenv`)
+
+```
+virtualenv mkdocs_env
+source mkdocs_env/bin/activate
+pip install mkdocs
+```
+
+- mkdocs in the nutshell
+
+```
+mkdocs build
+mkdocs gh-deploy
+```
+
+This will deploy your copy of `RNAsik` docs to your [github-pages (gh-pages)](https://pages.github.com/)
+You actually don't need to do that, you don't need deploy your own copy of the docs to your branch. Just use `mkdocs
+server` (read below) to prerview changes and send them through to me.
 
 ```
 git clone https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe
 cd RNAsik-pipe
-mkdocs serve
+# do docs changes
 ```
 
-This will give you live updates to you copy of the docs, default URL should be [localhost:8000](localhost:8000), but it will tell you that once you've started the server. Then simply use your favourite text editor to edit markdown documents. Commit your changes, don't be afraid to be verbose, say what you've added/changed/removed in your commit message. And send me PR
+- get localhost server (to preview your changes)
+
+```
+mkdocs server
+```
+
+This will give you live updates to you copy of the docs, default URL should be [localhost:8000](localhost:8000), but it will tell you that once you've started the server. Then simply use your favourite text editor to edit markdown documents. Commit your changes, don't be afraid to be verbose, say what you've added/changed/removed in your commit message. And send me PR.
 
 ## User's group
 
@@ -27,7 +52,7 @@ This will give you live updates to you copy of the docs, default URL should be [
 
 ## Developing pipeline further
 
-I need to write a more comprehensive developer guide at sometime soon. Any contributions are again extremely welcomed and again as I've mentioned in the [documentations](#documentations) section above, any contributions need to come through pull request (PR). 
+I need to write a more comprehensive developer guide at sometime soon. Any contributions are again extremely welcomed and again as I've mentioned in the [documentations](#documentations) section above, any contributions need to come through pull request (PR).
 
 To summarise briefly layouts of the `src/`:
 
@@ -90,6 +115,7 @@ conda install -y --use-local rnasik
     - set up an account at [Anacoda](anaconda.org)
     - `anaconda login`
     - `anaconda upload <path_to_file.tar.bz2>`
+    - `anaconda upload <path_to_file.tar.bz2>` --label dev
 
[…]
