### 2018-05-30 #583 @serine
@pjotrp @andrewyatz Let me guys know if you need anything else from me.

Just to reiterate that best way to install is to use `conda install`. I actually shifted myself and now use `conda install` all the time, cause I tend to jump between diff machines.

Also just to let you know that `RNAsik` have moved since I submitted for review. I'm now up to `1.5.1` version and making `1.5.2` soonish.. there have been some changes, but all in backward compatible manner.

You can only get `1.5.1` through conda install, but that shouldn't effect the review right?

Cheers
