### 2018-05-30 2f80598
Wrote more test, updated config and set all values that aren't yet used to -1 rather then zero. This is mainly because when you parseInt() and it's a string then return value is zero which is confusing since I want to catch if wrong value - string was set instead of int.
