+++ .travis.yml
@@ -27,4 +27,6 @@ install:
 script:
 
   - bds -test test/sikTests.bds
+  # next two should result in the same thing
+  - RNAsik
   - RNAsik --help
