### 2018-05-30 7807161
add mock refFiles directory into test/ dir. all are empty files I just need the file names and directory structure at this stage. Wrote more test Starting to form structure, made sikCommon.bds inside test/ dir. I think this is good idea? basically holding common variables in there Updated travis with new sik_test.bash script that runs all the tests now
