### 2018-06-15 daf507d
fixed bugs in get_gtf function
