### 2018-06-17 #583 @pjotrp
Hi @serine, sorry about the delay. So, you are saying that a conda install should work. Do you happen to have some example fastq files, i.e, something I can try? I am proposing I run it myself with your help. If it works I can switch to a reviewer role and I'll ask someone else to be the editor. It has happened before.
