### 2018-06-18 60630dd
Introduced short args notation to python scripts
