### 2018-06-18 6bda91c
Added testing for get_geneids.py script, this isn't python unit test rather simple "does the tool run" type test. With type will add different file types, i.e gtf, gff and saf to improve the testing.
Also tmp included bds install from dev conda channel, while waiting on
next release.
