### 2018-06-18 4c81b02
adding big ish test, full RNAsik run on yeast data
