### 2018-06-19 2e18d9e
wrote quick_install.bash script that meant to bootstrap miniconda and conda install rnasik. That script is now linked to quick start page in the docs. Updated travis to use that quick install script.
