+++ .travis.yml
@@ -6,15 +6,10 @@ language: bds
 
 before_install:
 
-  - if [[ ${TRAVIS_PYTHON_VERSION:0:1} == "2" ]]; then
-      wget http://repo.continuum.io/miniconda/Miniconda-2.0.0-Linux-x86_64.sh -O miniconda.sh;
-    else
-      wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh; fi
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH=$HOME/miniconda/bin:$PATH
+  - bash docs/extras/quick_install.bash
 
 install:
-
+  
   - conda config --add channels defaults
   - conda config --add channels conda-forge
   - conda config --add channels bioconda
