### 2018-06-19 2229d96
Moved quick_install.bash script into scripts, this is the right location for it. Also included hard coded url to the scripts so that the latest copy is available without needed to rebuild the site everytime
