### 2018-06-19 e63edcc
simplified travisCI script
