### 2018-06-19 c53adf6
qualimap has some issues with DISPLAY variable, explicitly unsetting it
