### 2018-06-19 #583 @serine
@pjotrp yes, conda installable. I just wrote [quick_install.bash](https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe/blob/master/scripts/quick_install.bash) script that obviously does quick install, but I'm also using it with my travisCI testing.

As you can see from [quick start section](https://monashbioinformaticsplatform.github.io/RNAsik-pipe/docs/#quick-start) in the docs that you can just run the script and export executables into your PATH. The details of the script are explained under [installation section](https://monashbioinformaticsplatform.github.io/RNAsik-pipe/docs/#installation) or just `cat` the script obviously

The actual test data again explained in [data set for testing section](https://monashbioinformaticsplatform.github.io/RNAsik-pipe/docs/#data-set-for-testing) and I've updated the README with the command to run, but here it is just in case

```
RNAsik -fqDir http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/fqFiles.txt \
       -align star \
       -fastaRef ftp://ftp.ensembl.org/pub/release-91/fasta/saccharomyces_cerevisiae/dna/Saccharomyces_cerevisiae.R64-1-1.dna_sm.toplevel.fa.gz \
       -gtfFile ftp://ftp.ensembl.org/pub/release-91/gtf/saccharomyces_cerevisiae/Saccharomyces_cerevisiae.R64-1-1.91.gtf.gz \
       -counts \
       -paired \
       -all
```

There are a couple of things I should mention.

- BigDataScript (bds) just moved into version 2 (yay) but I had to patch up my pipeline a bit, because of that I'm using specific commit of bds that works. Right now it is hard coded into [quick_install.bash](https://raw.githubusercontent.com/MonashBioinformaticsPlatform/RNAsik-pipe/master/scripts/quick_install.bash) script. Also note that `RNAsik` is under active development and version 1.5.1 has moved on fair bit with bug fixes and new feature. I once again hard coded RNAsik specific commit into the install script.

I'm waiting on bds to make next `v2.0rc9` release before making my next release `1.5.2`

I understand that there are a lot ifs here. But things are out of my hands with respect to bds development.

Basically if you just run the install script and test command I posted above it will work. Just note that STAR aligner needs fair bit of RAM, the test data set is small and I think 10-15 Gb of RAM should be plenty for mouse or human you really need ~ 32 Gb of RAM

thanks
