### 2018-06-19 #583 @serine
Forgot to mention new feature in `RNAsik` is ```-fqDir http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/fqFiles.txt```

where `-fqDir` can either take a list with files or directory. If text file is given, files expected to be one per line and if directory is give it will get traversed recursively looking for fastq files

the content of `fqFiles.txt` is

```
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/ctrl1_R1.fastq.gz
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/ctrl1_R2.fastq.gz
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/ctrl2_R1.fastq.gz
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/ctrl2_R2.fastq.gz
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/Ind-Z1_R1.fastq.gz
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/Ind-Z1_R2.fastq.gz
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/Ind-Z2_R1.fastq.gz
http://bioinformatics.erc.monash.edu/home/kirill/sikTestData/rawData/IndustrialAntifoamAgentsYeastRNAseqData/Ind-Z2_R2.fastq.gz
```

`RNAsik` will handle the download for you. Also note that `fqFiles.txt` is on url as well, again download and parsing are all handled internally, just run the command as is above.
