### 2018-06-19 fefcb85
updated install script with latest RNAsik version
