### 2018-07-05 5d3ab6e
updated docs in preparation for next release
