### 2018-07-05 5d7170b
updated quick_install script and checking update in the travisCI
