### 2018-07-10 32d71c4
updated docs and README
