### 2018-07-11 4589ac0
added tutorials tab to the docs, with some links to useful materials
