+++ docs/tutorials.md
@@ -1,0 +1,9 @@
+# Tutorials
+
+So far I don't have any tutorials that I've wrote specific to this (RNAseq) topic, but
+here is few very useful links to learning resource.
+
+- [Hands-on Training in RNA-Seq Data Analysis](https://legacy.gitbook.com/book/ycl6/rna-seq-data-analysis/details)
+- [Applied Computational Genomics Course](https://github.com/quinlan-lab/applied-computational-genomics)
+- [Bioconductor RNAseq tutorial](https://master.bioconductor.org/packages/release/workflows/vignettes/rnaseqGene/inst/doc/rnaseqGene.html)
+- [RNA-seqlopedia](https://rnaseq.uoregon.edu/)
