+++ mkdocs.yml
@@ -7,6 +7,7 @@ edit_uri: ""
 pages:
   - About: index.md
   - Docs: docs.md
+  - Tutes: tutorials.md
   - Help: help.md
   - Contributing: contrib.md
   - Roadmap: roadmap.md
