### 2018-07-16 #583 @serine
great, thanks. 1.5.2 is out now by the way
