### 2018-07-17 #583 @pjotrp
I installed the software with bioconda bootstrapping from the GNU Guix conda package. Nice work! Mostly worked out of the box. I'll show a writeup when I am done.

@serine - do you have a test dataset I can just download and run? Sorry for the pain, but I don't want to be hunting down NCBI resources. Even a small set will do.
