### 2018-07-18 #583 @pjotrp
My report: I can report the software installs and runs to satisfaction. In fact, kudos @serine and colleagues for creating such an advanced pipeline which can be run on almost any Linux system out of the box with nice reporting output. I hope people will start using it. Since it is bionconda based you should also look at making it part of the common workflow initiative https://www.commonwl.org/ with @mr-c.

Let me tick the relevant boxes which means we can process the publication for acceptance. @serine, sorry about the delays, it is not the usual, but then your software is not the usual either :). Good stuff.
