### 2018-07-18 #583 @serine
This is great, thanks heaps !

About the second part; I've done Zenodo here https://zenodo.org/record/1315013#.W0_PY3VuaV4

[![DOI](https://zenodo.org/badge/DOI/10.21105/joss.00583.svg)](https://doi.org/10.21105/joss.00583)

I was a bit confused about which DOI to use.. I just tool whatever was there in the pds that you've generated.

About the first bit _can you confirm that all references in your bibliography have a_

Looking through the references these tools don't have doi

```
Broadinstitute. n.d. “Picard-Tools.” broadinstitute. http://broadinstitute.github.io/
picard.
```
```
Li, Heng. 2013. “Aligning Sequence Reads, Clone Sequences and Assembly Contigs with
BWA-MEM,” March. http://arxiv.org/abs/1303.3997.
```
```
Powell, David. 2015. “Degust: Powerfull and User Friendly Front-End Data Analsysis,
Visualisation and Exploratory Tool for Rna-Sequencing.” github. http://degust.erc.
monash.edu
```

do I need to do something about this?
