### 2018-07-19 80ecedd
updated docs, removed advanced tab - not that interesting, added checklist tab, checklist for RNAseq analysis
