### 2018-07-19 #583 @mr-c
Would an RRID URI be acceptable? I like to cite both the idea of a piece of software and any relevant methods papers.

Picard:

https://identifiers.org/rrid/RRID:SCR_006525

Degust:

https://identifiers.org/rrid/RRID:SCR_001878

For bwa-mem I suggest listing  both the Arxiv paper and the RRID
https://arxiv.org/abs/1303.3997
https://identifiers.org/rrid/RRID:SCR_010910

(There are citation instructions for BWA but it doesn't mention the `mem` algorithm: http://bio-bwa.sourceforge.net/ just `short` and  `sw`)
