### 2018-07-19 #583 @serine
@mr-c was that comment _Would an RRID URI be acceptable?_ to me or at @arfon cause I'm happy with what ever works
