### 2018-07-19 #583 @serine
@mr-c how about now https://zenodo.org/record/1315013#.W1A5FHVuaV4 ?

using

```
pandoc -o check.pdf --bibliography=paper.bib paper.md
```
builds to this

[check.pdf](https://github.com/openjournals/joss-reviews/files/2208689/check.pdf)

not sure how it'll look with JOSS compilation.. also don't think I can do much about footer overlap

if you think this looks ok I can commit this change to master and perhaps ask @pjotrp to rebuild final pdf if that's ok?
