### 2018-07-19 d567e44
made paper.md more readable, thanks to @mr-c for input
