+++ paper.md
@@ -28,8 +28,13 @@ bibliography: paper.bib
 # Summary
 
 RNA sequencing (RNA-seq) is one of many applications of high throughput sequencing, in which millions of short sequence reads, typically around 100 bases long, are produced from RNA samples with the aim of characterising entire transcriptomes. In order to analyse RNA-seq data, multiple bioinformatics tools are collected together into a pipeline, in which each tool accepts processed data from the previous tool as its input. The RNAsik pipeline streamlines processing of RNA-seq data and facilitates the production of reproducible results. This pipeline can run standalone on workstations, cloud instances, or on High Performance Computing (HPC) clusters. A single RNAsik run gives a comprehensive overview of the experiment and produces output suitable for Differential Gene Expression (DGE) analysis.
-With an alignment based approach, RNAsik incorporates two main steps: 1) alignment of short reads from FASTQ files to a reference genome or transcriptome; and 2) counting the number of reads mapped to annotated genomic features (such as genes). The table of counts generated by RNAsik can be further analysed with any contemporary count-based DGE tools/packages - one such package particularly suited to the task is Degust [@Powell2015], a powerful front-end and user friendly tool for DGE data analysis, visualisation and exploration. Additionally RNAsik can produce a number of different quality control (QC) metrics, such as sequencing quality metrics reported using FastQC [@Bioinformatics2011-ve], intra- and inter-genic mapping rates estimation using QualiMap [@Okonechnikov2016-jk], and sequencing library size and GC bias estimation using Picard Tools [@picard-tools2018]. These QC metrics are automatically summarised into a single, dynamic, HTML report generated by MulitQC [@Ewels2016-ik]. Other features of the RNAsik pipeline include the ability to mark duplicated reads using Picard Tools [@picard-tools2018], sorting and indexing of alignments using Samtools [@Li2009-aa] to enable viewing in genome browser applications such as IGV [Robinson2011-du], an enhanced table of counts with additional meta-information about each gene (e.g. biotype and human readable gene names), and ready-to-use coverage plots for every sample using bedtools2 [@Quinlan2010-am] and UCSC tools [@Raney2014-ti]. The RNAsik pipeline logs every step of processing including the number of samples and associated FASTQ files, software tool versions and sequencing strand information. 
+
+With an alignment based approach, RNAsik incorporates two main steps: 1) alignment of short reads from FASTQ files to a reference genome or transcriptome; and 2) counting the number of reads mapped to annotated genomic features (such as genes). The table of counts generated by RNAsik can be further analysed with any contemporary count-based DGE tools/packages - one such package particularly suited to the task is Degust [@Powell2015], a powerful front-end and user friendly tool for DGE data analysis, visualisation and exploration. Additionally RNAsik can produce a number of different quality control (QC) metrics, such as sequencing quality metrics reported using FastQC [@Bioinformatics2011-ve], intra- and inter-genic mapping rates estimation using QualiMap [@Okonechnikov2016-jk], and sequencing library size and GC bias estimation using Picard Tools [@picard-tools2018]. These QC metrics are automatically summarised into a single, dynamic, HTML report generated by MulitQC [@Ewels2016-ik].
+
[…]
