### 2018-08-07 #583 @arfon
@serine - you appear to have associated the JOSS DOI with your archive in Zenodo (https://zenodo.org/record/1315013#.W2m3Nv5Khjt). Could you please follow [this guide](https://guides.github.com/activities/citable-code/) and create an archive of the GitHub repository (https://github.com/MonashBioinformaticsPlatform/RNAsik-pipe) that has a Zenodo DOI and not a JOSS one?
