### 2018-08-08 #583 @serine
right, thanks for the link to the guide @arfon . I've done everything but the new release. I'll have to do a bit a work for my next release. I can't just slap new release tag there is a [timeline and things](https://monashbioinformaticsplatform.github.io/RNAsik-pipe/roadmap/) :)
Will try to get it done in the next couple of days.

Pretty exciting to be so close !

Thanks
