### 2018-08-27 #583 @serine
@pjotrp sorry about the wait, was away for a week and was playing catch up. I just made another release that is hooked to zenodo. https://zenodo.org/account/settings/github/repository/MonashBioinformaticsPlatform/RNAsik-pipe

I didn't realise before that I could have "annotated" any of my previous releases and that would have showed up in zenodo. I thought that after setting up a webhook zenodo was "waiting" for an event (next release). Anyway I think all done now.

@arfon is there anything else I need to do?

Thanks
